Refactored code to lib.rs
